requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
filelock==3.9.0
frozenlist==1.3.3
grpcio==1.51.3
grpcio-tools==1.51.3
idna==3.4
lxml==4.9.2
macholib==1.16.2
multidict==6.0.4
openai==0.27.0
protobuf==4.22.0
pycryptodomex==3.17
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
regex==2022.10.31
requests==2.28.2
tiktoken==0.3.0
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
